+++ database/migrations/2023_02_08_171000_create_working_mode_table.php
@@ -1,0 +1,42 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+
+    private string $table_name = "working_mode";
+
+    private function getTableName(): string
+    {
+        return $this->table_name;
+    }
+
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create($this->getTableName(), function (Blueprint $table) {
+            $table->id();
+            $table->string('name');
+            $table->integer('view_order');
+            $table->timestampsTz();
+            $table->softDeletesTz();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists($this->getTableName());
+    }
+};