+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next version
 + Refactor Semaphore (old docker-compose)
 + Add Qodana
++ Add working mode (migration, repository, gateway and controller)
 
 ## 0.10.0
 + Add tools and techonologies (migration, repository, gateway and controller)